修改主键泛型为Int类型

+++ src/main/java/com/lord/advertis/dao/ModuleDao.java
@@ -3,6 +3,6 @@ package com.lord.advertis.dao;
 import com.lord.advertis.base.BaseMapper;
 import com.lord.advertis.domain.Module;
 
-public interface ModuleDao extends BaseMapper<Module, Long>{
+public interface ModuleDao extends BaseMapper<Module, Integer>{
 
 }
